lab/gcp/5-11-2025-mensa/func_stat/requirements.txt
google-cloud-firestore==2.22.0

# Se contiene una web application 
flask==2.3.3
gunicorn==23.0.0